mmc: sd: Set card state to removed upon ejection
When the SD card is removed there's a detect
work that may or may not run immediately.
In the meantime if a request is issued, then
runtime_resume would pass without actually resuming
the card. Runtime-suspend turns-off the clocks too.

Thus writing to controller registers without resuming
causes the NOC error.

Fix this by setting the appropriate card state in resume.

CRs-fixed: 2452075
Change-Id: I70737c2b355ccf7a7c5e89130d57b458c1a7365b
Signed-off-by: Asutosh Das <asutoshd@codeaurora.org>